Refactor GraphVisual layout code
The GraphVisual does not compute the locations of the nodes
anymore. This is delegated to the `layout` property.

A layout is any iterator yielding a (node_vertices, line_vertices)
tuple. Using the iterator we can let layout algorithms yield
intermediate results, which the GraphVisual can use to animate
the layout.

The user can specify if it wants the the application of a layout on
a graph to be animated. But it is the responsibility of the user to
call GraphVisual.animate_layout() repeatedly until the calculation
of the layout is done.